AI_SERVICE/requirements.txt
tensorflow
flask
numpy
pillow
flask-cors
